Agent_UI_0.2/source_code/backend/requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
sentence-transformers
torch
transformers
pillow
pytesseract
pdfplumber
python-docx
python-pptx
opencv-python

